@@ -3,24 +3,23 @@ import 'dart:io';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:mask_text_input_formatter/mask_text_input_formatter.dart';
-import 'package:flutter_svg/flutter_svg.dart';
+//import 'package:flutter_svg/flutter_svg.dart';
 import 'package:myevpanet/main.dart';
 import 'package:myevpanet/main_screen/main_widget.dart';
 import 'package:responsive_flutter/responsive_flutter.dart';
 import 'package:myevpanet/api/api.dart';
 import 'package:fluttertoast/fluttertoast.dart';
 import 'package:flutter/services.dart';
-import 'package:myevpanet/webview_screens/order_widget.dart';
+//import 'package:myevpanet/webview_screens/order_widget.dart';
 
 class LoginWidget extends StatefulWidget {
   @override
   State<StatefulWidget> createState() => LoginWidgetState();
 }
 
-class LoginWidgetState extends State with SingleTickerProviderStateMixin{
-
+class LoginWidgetState extends State with SingleTickerProviderStateMixin {
   @override
-  void initState(){
+  void initState() {
     super.initState();
     currentGuidIndex = 0;
     SystemChrome.setPreferredOrientations([
@@ -29,93 +28,97 @@ class LoginWidgetState extends State with SingleTickerProviderStateMixin{
     ]);
   }
 
-  final String assetName = 'assets/images/evpanet_auth_logo.svg';
+  //final String assetName = 'assets/images/evpanet_auth_logo.svg';
+  final String assetName = 'assets/images/splash_logo.png';
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      resizeToAvoidBottomPadding: false,
+      resizeToAvoidBottomInset: false,
       body: Stack(
         fit: StackFit.expand,
         children: <Widget>[
-        Container(
-          decoration: BoxDecoration(
-            gradient: LinearGradient(
+          Container(
+            decoration: BoxDecoration(
+                gradient: LinearGradient(
+              colors: [Color(0xff11273c), Color(0xff3c5d7c)],
               begin: Alignment.topLeft,
               end: Alignment.bottomRight,
-              colors: [
-                Color(0xff11273c),
-                Color(0xff3c5d7c)
-              ]
-            )
-          ),
-          child: Center(
-            child: SingleChildScrollView(
-              child: ConstrainedBox(
-              constraints: BoxConstraints(maxHeight: MediaQuery.of(context).size.height),
-                child: Container(
-                  padding: EdgeInsets.only(
-                    top: ResponsiveFlutter.of(context).verticalScale(20),
-                  ),
-                  height: ResponsiveFlutter.of(context).verticalScale(600),
-                  width: ResponsiveFlutter.of(context).scale(300),
-                  child: Column(
-                    mainAxisAlignment: MainAxisAlignment.start,
-                    children: <Widget>[
-                      _buildLogoTop(),
-                      _buildPhoneField(),
-                      _buildUIDField(),
-                      _buildSubmitButton(),
-                      Padding(
-                        padding: EdgeInsets.symmetric(horizontal: 10.0, vertical: 5.0),
-                        child: LinearProgressIndicator(
-                            value: currentGuidIndex / (guids.isNotEmpty ? guids.length : 1),
-                            backgroundColor: Color(0xff3c5d7c),
-                            valueColor: AlwaysStoppedAnimation<Color>(Colors.white)
+            )),
+            child: Center(
+              child: SingleChildScrollView(
+                child: ConstrainedBox(
+                  constraints: BoxConstraints(
+                      maxHeight: MediaQuery.of(context).size.height),
+                  child: Container(
+                    padding: EdgeInsets.only(
+                      top: ResponsiveFlutter.of(context).verticalScale(20),
+                    ),
+                    height: ResponsiveFlutter.of(context).verticalScale(600),
+                    width: ResponsiveFlutter.of(context).scale(300),
+                    child: Column(
+                      mainAxisAlignment: MainAxisAlignment.start,
+                      children: <Widget>[
+                        _buildLogoTop(),
+                        _buildPhoneField(),
+                        _buildUIDField(),
+                        _buildSubmitButton(),
+                        Padding(
+                          padding: EdgeInsets.symmetric(
+                              horizontal: 10.0, vertical: 5.0),
+                          child: LinearProgressIndicator(
+                              value: currentGuidIndex /
+                                  (guids.isNotEmpty ? guids.length : 1),
+                              backgroundColor: Color(0xff3c5d7c),
+                              valueColor:
+                                  AlwaysStoppedAnimation<Color>(Colors.white)),
                         ),
-                      ),
-                      Container(
-                        padding: EdgeInsets.only(
-                          top: ResponsiveFlutter.of(context).moderateScale(8),
-                          bottom: ResponsiveFlutter.of(context).moderateScale(8),
-                        ),
-                        child: Row(children: <Widget>[
-                          Expanded(
-                            child: new Container(
-                                margin: const EdgeInsets.only(left: 10.0, right: 15.0),
-                                child: Divider(
-                                  color: Color(0xffd3edff),
-                                  height: 50,
-                                )),
+                        /*
+                        Container(
+                          padding: EdgeInsets.only(
+                            top: ResponsiveFlutter.of(context).moderateScale(8),
+                            bottom:
+                                ResponsiveFlutter.of(context).moderateScale(8),
                           ),
-                          Text(
-                            "или",
-                            style: TextStyle(
-                                color: Color(0xffd3edff),
-                              fontSize: ResponsiveFlutter.of(context).fontSize(1.5)
+                          child: Row(children: <Widget>[
+                            Expanded(
+                              child: new Container(
+                                  margin: const EdgeInsets.only(
+                                      left: 10.0, right: 15.0),
+                                  child: Divider(
+                                    color: Color(0xffd3edff),
+                                    height: 50,
+                                  )),
                             ),
-                          ),
-                          Expanded(
-                            child: new Container(
-                                margin: const EdgeInsets.only(left: 15.0, right: 10.0),
-                                child: Divider(
+                            Text(
+                              "или",
+                              style: TextStyle(
                                   color: Color(0xffd3edff),
-                                  height: 50,
-                                )),
-                          ),
-                        ]),
-                      ),
-                      _buildConnectRequestButton(),
-                    ],
+                                  fontSize: ResponsiveFlutter.of(context)
+                                      .fontSize(1.5)),
+                            ),
+                            Expanded(
+                              child: new Container(
+                                  margin: const EdgeInsets.only(
+                                      left: 15.0, right: 10.0),
+                                  child: Divider(
+                                    color: Color(0xffd3edff),
+                                    height: 50,
+                                  )),
+                            ),
+                          ]),
+                        ),
+                        _buildConnectRequestButton(),*/
+                      ],
+                    ),
+                    alignment: Alignment(0.0, 0.0),
                   ),
-                  alignment: Alignment(0.0, 0.0),
                 ),
               ),
             ),
           ),
-        ),
-      ],
-    ),
+        ],
+      ),
     );
   }
 
@@ -131,17 +134,18 @@ class LoginWidgetState extends State with SingleTickerProviderStateMixin{
       //номер телефона достаточной длины и ИД тоже корректный можно идти дальше
       var phone = int.parse(_phoneNumber);
       var id = _userID;
-      if (verbose >=1) print('Entered phone number is: $phone');
-      if (verbose >=1) print('Entered ID is: $id');
+      if (verbose >= 1) print('Entered phone number is: $phone');
+      if (verbose >= 1) print('Entered ID is: $id');
 
       //requesting auth from server
-      Map<String, String> result = await RestAPI().authorizeUserPOST('+' + phone.toString(), id, devKey);
+      Map<String, String> result =
+          await RestAPI().authorizeUserPOST('+' + phone.toString(), id, devKey);
 
       //обработка ответа сервера
       bool _checks = true;
       if (result['answer'] == 'isFull') {
         var _body = json.decode(result['body']);
-        if (verbose >=1) print(_body);
+        if (verbose >= 1) print(_body);
         if (!_body['error']) {
           //ошибок нет. в ответе лежат гуиды
           if (registrationMode == 'new') {
@@ -158,69 +162,70 @@ class LoginWidgetState extends State with SingleTickerProviderStateMixin{
           _checks = false;
           if (verbose >= 1) print('Got Error: ${_body['message']}');
           Fluttertoast.showToast(
-            msg: "Абонент(ов) с такими данными не найдено",
-            toastLength: Toast.LENGTH_SHORT,
-            gravity: ToastGravity.CENTER,
-            timeInSecForIosWeb: 1,
-            backgroundColor: Colors.red,
-            textColor: Colors.white,
-            fontSize: 16.0
-          );
+              msg: "Абонент(ов) с такими данными не найдено",
+              toastLength: Toast.LENGTH_SHORT,
+              gravity: ToastGravity.CENTER,
+              timeInSecForIosWeb: 1,
+              backgroundColor: Colors.red,
+              textColor: Colors.white,
+              fontSize: 16.0);
         }
       } else {
         //ответ был не получен из-за ошибки сети
         _checks = false;
-        if (verbose >=1) print('network error :(');
+        if (verbose >= 1) print('network error :(');
       }
       if (_checks) {
-        if (verbose >=1) print('Got good GUID(s)');
+        if (verbose >= 1) print('Got good GUID(s)');
         //fbHelper.saveDeviceToken(result);
         //сохранение списка гуидов в файл
-        if (verbose >=1) print(guids);
+        if (verbose >= 1) print(guids);
         final _guidsfile = await FileStorage('guidlist.dat').localFile;
-        _guidsfile.writeAsString(jsonEncode(guids), mode: FileMode.write, encoding: utf8);
+        _guidsfile.writeAsString(jsonEncode(guids),
+            mode: FileMode.write, encoding: utf8);
         currentGuidIndex = 0;
         //запрашиваем сервер данные всех пользователей и сохраняем (обновляем) файлы
         for (var guid in guids) {
           var usersRequest = await RestAPI().userDataGet(guid, devKey);
-          if (verbose >=1) print('$guid: $usersRequest');
+          if (verbose >= 1) print('$guid: $usersRequest');
           currentGuidIndex++;
           setState(() {});
         }
         currentGuidIndex = 0;
-        Navigator.of(context).pushReplacement(MaterialPageRoute(builder: (BuildContext context) => MainScreenWidget()));
+        Navigator.of(context).pushReplacement(MaterialPageRoute(
+            builder: (BuildContext context) => MainScreenWidget()));
       }
     } else {
       Fluttertoast.showToast(
-        msg: "Введены некорректные данные",
-        toastLength: Toast.LENGTH_SHORT,
-        gravity: ToastGravity.CENTER,
-        timeInSecForIosWeb: 1,
-        backgroundColor: Colors.red,
-        textColor: Colors.white,
-        fontSize: 16.0
-      );
+          msg: "Введены некорректные данные",
+          toastLength: Toast.LENGTH_SHORT,
+          gravity: ToastGravity.CENTER,
+          timeInSecForIosWeb: 1,
+          backgroundColor: Colors.red,
+          textColor: Colors.white,
+          fontSize: 16.0);
     }
   }
-  Widget _buildLogoTop(){
+
+  Widget _buildLogoTop() {
     return Center(
-      child: Container(
-          padding: EdgeInsets.only(
-            top: ResponsiveFlutter.of(context).moderateScale(20),
-            bottom: ResponsiveFlutter.of(context).moderateScale(20),
-            right: ResponsiveFlutter.of(context).moderateScale(8),
-            left: ResponsiveFlutter.of(context).moderateScale(8),
-          ),
-          child: SvgPicture.asset(
-            assetName,
-            color: Color(0xffd3edff),
-          )
-      )
-    );
+        child: Container(
+            padding: EdgeInsets.only(
+              top: ResponsiveFlutter.of(context).moderateScale(20),
+              bottom: ResponsiveFlutter.of(context).moderateScale(20),
+              right: ResponsiveFlutter.of(context).moderateScale(8),
+              left: ResponsiveFlutter.of(context).moderateScale(8),
+            ),
+            child: Image.asset(
+              assetName,
+              color: Color(0xffd3edff),
+            )));
   }
 
-  var phone = new MaskTextInputFormatter(mask: '+# (###) ###-##-##', filter: { "#": RegExp(r'[0-9]') });
-  var uid = new MaskTextInputFormatter(mask: '#####', filter: { "#": RegExp(r'[0-9]') });
+  var phone = new MaskTextInputFormatter(
+      mask: '+# (###) ###-##-##', filter: {"#": RegExp(r'[0-9]')});
+  var uid = new MaskTextInputFormatter(
+      mask: '#####', filter: {"#": RegExp(r'[0-9]')});
 
   Widget _buildPhoneField() {
     return Padding(
@@ -232,10 +237,7 @@ class LoginWidgetState extends State with SingleTickerProviderStateMixin{
           //controller: textEditingController,
           inputFormatters: [phone],
           keyboardType: TextInputType.phone,
-          style: TextStyle(
-            color: Color(0xffd3edff),
-            fontSize: 18.0
-          ),
+          style: TextStyle(color: Color(0xffd3edff), fontSize: 18.0),
           textCapitalization: TextCapitalization.characters,
           decoration: InputDecoration(
               labelText: 'Номер телефона',
@@ -244,18 +246,10 @@ class LoginWidgetState extends State with SingleTickerProviderStateMixin{
                 letterSpacing: 1,
               ),
               focusedBorder: UnderlineInputBorder(
-                borderSide: BorderSide(
-                    color: Color(0xffd3edff)
-                )
-              ),
+                  borderSide: BorderSide(color: Color(0xffd3edff))),
               enabledBorder: UnderlineInputBorder(
-                borderSide: BorderSide(
-                    color: Color(0xffd3edff)
-                )
-              )
-          ),
-        )
-    );
+                  borderSide: BorderSide(color: Color(0xffd3edff)))),
+        ));
   }
 
   Widget _buildUIDField() {
@@ -273,40 +267,37 @@ class LoginWidgetState extends State with SingleTickerProviderStateMixin{
           decoration: InputDecoration(
               labelText: 'Ваш ИД (ID)',
               labelStyle: TextStyle(
-                  color: Color(0xffd3edff),
-                  letterSpacing: 1,
+                color: Color(0xffd3edff),
+                letterSpacing: 1,
               ),
               focusedBorder: UnderlineInputBorder(
-                  borderSide: BorderSide(
-                      color: Color(0xffd3edff)
-                  )
-              ),
+                  borderSide: BorderSide(color: Color(0xffd3edff))),
               enabledBorder: UnderlineInputBorder(
-                  borderSide: BorderSide(
-                      color: Color(0xffd3edff)
-                  )
-              )
-          ),
-        )
-    );
+                  borderSide: BorderSide(color: Color(0xffd3edff)))),
+        ));
   }
 
   Widget _buildSubmitButton() {
-    return
-    Padding(
+    return Padding(
       padding: EdgeInsets.only(
         top: ResponsiveFlutter.of(context).moderateScale(20),
         bottom: ResponsiveFlutter.of(context).moderateScale(8),
         right: ResponsiveFlutter.of(context).moderateScale(8),
         left: ResponsiveFlutter.of(context).moderateScale(8),
       ),
-      child: RaisedButton(
-        shape: RoundedRectangleBorder(side: BorderSide(color: Color(0xff95abbf))),
-        elevation: 0.0,
+      child: ElevatedButton(
+        style: ElevatedButton.styleFrom(
+            primary: Color(0x858eaac2),
+            elevation: 0.0,
+            shape: RoundedRectangleBorder(
+                side: BorderSide(color: Color(0xff95abbf)))),
+        //shape: RoundedRectangleBorder(side: BorderSide(color: Color(0xff95abbf))),
+        //elevation: 0.0,
         onPressed: devKey != null ? authButtonPressed : null,
-        color: Color(0x858eaac2),
+        //color: Color(0x858eaac2),
         child: Padding(
-          padding: EdgeInsets.all(ResponsiveFlutter.of(context).moderateScale(16)),
+          padding:
+              EdgeInsets.all(ResponsiveFlutter.of(context).moderateScale(16)),
           child: Row(
             mainAxisAlignment: MainAxisAlignment.center,
             children: <Widget>[
@@ -326,7 +317,7 @@ class LoginWidgetState extends State with SingleTickerProviderStateMixin{
           ),
         ),
       ),
-      );
+    );
   }
 
   String getAuthButtonPressed() {
@@ -343,45 +334,47 @@ class LoginWidgetState extends State with SingleTickerProviderStateMixin{
     }
   }
 
+  /*
   Widget _buildConnectRequestButton() {
-    return
-      Padding(
-        padding: EdgeInsets.only(
-          top: ResponsiveFlutter.of(context).moderateScale(2),
-          bottom: ResponsiveFlutter.of(context).moderateScale(8),
-          right: ResponsiveFlutter.of(context).moderateScale(8),
-          left: ResponsiveFlutter.of(context).moderateScale(8),
-        ),
-        child: RaisedButton(
-          elevation: 0.0,
-          onPressed: devKey != null ? () {
-            // Тут колбасим открытие новой страницы с возможностью оставить заявку на подключение (Используем новый API).
-            Navigator.of(context).push(MaterialPageRoute(builder: (BuildContext context) => OrderView()));
-          } : null,
-          color: Color(0x408eaac2),
-          child: Padding(
-            padding: EdgeInsets.all(ResponsiveFlutter.of(context).moderateScale(8)),
-            child: Row(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: <Widget>[
-                Icon(
-                  Icons.person_add,
+    return Padding(
+      padding: EdgeInsets.only(
+        top: ResponsiveFlutter.of(context).moderateScale(2),
+        bottom: ResponsiveFlutter.of(context).moderateScale(8),
+        right: ResponsiveFlutter.of(context).moderateScale(8),
+        left: ResponsiveFlutter.of(context).moderateScale(8),
+      ),
+      child: RaisedButton(
+        elevation: 0.0,
+        onPressed: devKey != null
+            ? () {
+                // Тут колбасим открытие новой страницы с возможностью оставить заявку на подключение (Используем новый API).
+                Navigator.of(context).push(MaterialPageRoute(
+                    builder: (BuildContext context) => OrderView()));
+              }
+            : null,
+        color: Color(0x408eaac2),
+        child: Padding(
+          padding:
+              EdgeInsets.all(ResponsiveFlutter.of(context).moderateScale(8)),
+          child: Row(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: <Widget>[
+              Icon(
+                Icons.person_add,
+                color: Colors.white,
+              ),
+              Text(
+                devKey != null ? ' Подключиться' : 'Нет доступа к интернету',
+                style: TextStyle(
+                  fontSize: ResponsiveFlutter.of(context).fontSize(2),
+                  fontWeight: FontWeight.w700,
                   color: Colors.white,
                 ),
-                Text(
-                  devKey != null ? ' Подключиться' : 'Нет доступа к интернету',
-                  style: TextStyle(
-                    fontSize: ResponsiveFlutter.of(context).fontSize(2),
-                    fontWeight: FontWeight.w700,
-                    color: Colors.white,
-                  ),
-                ),
-              ],
-            ),
+              ),
+            ],
           ),
         ),
-      );
-  }
-
+      ),
+    );
+  }*/
 }
-
