@@ -7,111 +7,104 @@ import 'package:myevpanet/main_screen/setups.dart';
 String sendText;
 
 Widget blueZone_1(Map userInfo) {
-  return
-    Container(
-      color: Colors.blue,
-      height: 230.0,
-      child: Padding(
-        padding: EdgeInsets.only(
-          top: 10.0,
-          right: 20.0,
-          left: 20.0,
-          bottom: 20.0
-        ),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.stretch,
-          children: <Widget>[
-            Expanded(
-              flex: 1,
-              child: Container(
-                padding: EdgeInsets.only(
-                    top: 10.0
-                ),
-                  child: Row(
-                    //crossAxisAlignment: CrossAxisAlignment.stretch,
-                    children: <Widget>[
-                      Expanded(
-                        flex: 3,
-                        child: Container(
-                          padding: EdgeInsets.only(
-                              right: 30.0
-                          ),
-                          child: Column(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: <Widget>[
-                              Container(
-                                child: Text(
-                                  "Текущий тарифный план",
-                                  style: TextStyle(
-                                    color: Colors.white,
-                                    fontWeight: FontWeight.bold,
-                                  ),
-                                ),
-                              ),
-                              Container(
-                                padding: EdgeInsets.only(
-                                  top: 5.0
-                                ),
-                                child: Text(
-                                  userInfo["tarif_name"] + " (" + userInfo["tarif_sum"].toString() + " р.)",
-                                  style: TextStyle(
-                                    color: Colors.white,
-                                    fontSize: 18,
-                                  ),
-                                ),
-                              ),
-                            ],
-                          ),
-                        ),
-                      ),
-                      Expanded(
-                        flex: 1,
-                          child: Container(),
-                      ),
-                    //Text('ID: ${userInfo['id'].toString()}'),
-                    //Text('Тарифный план: ' + userInfo["tarif_name"] + " (" + userInfo["tarif_sum"]?.toString() + "р.)" ),
-                    ],
-                  )
-              ),
-            ),
-            Expanded(
-              flex: 2,
-              child: Container(
+  return Container(
+    color: Colors.blue,
+    height: 230.0,
+    child: Padding(
+      padding:
+          EdgeInsets.only(top: 10.0, right: 20.0, left: 20.0, bottom: 20.0),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.stretch,
+        children: <Widget>[
+          Expanded(
+            flex: 1,
+            child: Container(
                 padding: EdgeInsets.only(top: 10.0),
                 child: Row(
+                  //crossAxisAlignment: CrossAxisAlignment.stretch,
                   children: <Widget>[
                     Expanded(
-                    flex: 1,
-                      child: Padding(
-                        padding: EdgeInsets.only(
-                        //top: 20.0
+                      flex: 3,
+                      child: Container(
+                        padding: EdgeInsets.only(right: 30.0),
+                        child: Column(
+                          crossAxisAlignment: CrossAxisAlignment.start,
+                          children: <Widget>[
+                            Container(
+                              child: Text(
+                                "Текущий тарифный план",
+                                style: TextStyle(
+                                  color: Colors.white,
+                                  fontWeight: FontWeight.bold,
+                                ),
+                              ),
+                            ),
+                            Container(
+                              padding: EdgeInsets.only(top: 5.0),
+                              child: Text(
+                                userInfo["tarif_name"] +
+                                    " (" +
+                                    userInfo["tarif_sum"].toString() +
+                                    " р.)",
+                                style: TextStyle(
+                                  color: Colors.white,
+                                  fontSize: 18,
+                                ),
+                              ),
+                            ),
+                          ],
                         ),
-                          child: SizedBox(
+                      ),
+                    ),
+                    Expanded(
+                      flex: 1,
+                      child: Container(),
+                    ),
+                    //Text('ID: ${userInfo['id'].toString()}'),
+                    //Text('Тарифный план: ' + userInfo["tarif_name"] + " (" + userInfo["tarif_sum"]?.toString() + "р.)" ),
+                  ],
+                )),
+          ),
+          Expanded(
+            flex: 2,
+            child: Container(
+              padding: EdgeInsets.only(top: 10.0),
+              child: Row(
+                children: <Widget>[
+                  Expanded(
+                    flex: 1,
+                    child: Padding(
+                      padding: EdgeInsets.only(
+                          //top: 20.0
+                          ),
+                      child: SizedBox(
                         /*height: 165,
                           width: 165,*/
-                            child: Stack(
-                              children: <Widget>[
-                                SizedBox(
-                                  child: CircularProgressIndicator(
+                        child: Stack(
+                          children: <Widget>[
+                            SizedBox(
+                              child: CircularProgressIndicator(
                                 value: 1,
-                                valueColor: new AlwaysStoppedAnimation(Colors.white),
+                                valueColor:
+                                    new AlwaysStoppedAnimation(Colors.white),
                                 strokeWidth: 7.0,
                               ),
                               height: 100.0,
                               width: 100.0,
                             ),
-                              SizedBox(
-                                child: CircularProgressIndicator(
-                                value: secToDate(userInfo['packet_secs'])/30,
+                            SizedBox(
+                              child: CircularProgressIndicator(
+                                value: secToDate(userInfo['packet_secs']) / 30,
                                 //value: 0.875,
-                                valueColor: new AlwaysStoppedAnimation(Colors.lightBlueAccent),
+                                valueColor: new AlwaysStoppedAnimation(
+                                    Colors.lightBlueAccent),
                                 strokeWidth: 5.0,
                               ),
                               height: 100.0,
                               width: 100.0,
                             ),
-                              SizedBox(
-                                child: Column(
+                            SizedBox(
+                              child: Column(
                                 mainAxisAlignment: MainAxisAlignment.center,
                                 crossAxisAlignment: CrossAxisAlignment.center,
                                 children: <Widget>[
@@ -134,8 +127,7 @@ Widget blueZone_1(Map userInfo) {
                                                 foreground: Paint()
                                                   ..style = PaintingStyle.stroke
                                                   ..color = Colors.white70
-                                                  ..strokeWidth = 1
-                                            ),
+                                                  ..strokeWidth = 1),
                                           ),
                                           Text(
                                             "${secToDate(userInfo['packet_secs']).truncate()}",
@@ -155,7 +147,6 @@ Widget blueZone_1(Map userInfo) {
                                       ),
                                     ],
                                   ),
-
                                 ],
                               ),
                               height: 100.0,
@@ -164,10 +155,11 @@ Widget blueZone_1(Map userInfo) {
                           ],
                         ),
                       ),
-                    ),),
-                    Expanded(
-                      flex: 2,
-                      child: Row(
+                    ),
+                  ),
+                  Expanded(
+                    flex: 2,
+                    child: Row(
                       crossAxisAlignment: CrossAxisAlignment.end,
                       mainAxisAlignment: MainAxisAlignment.end,
                       children: <Widget>[
@@ -175,11 +167,9 @@ Widget blueZone_1(Map userInfo) {
                           padding: EdgeInsets.only(top: 10.0),
                         ),
                         Padding(
-                          padding: EdgeInsets.only(
-                              bottom: 6.0,
-                              right: 4.0
-                          ),
-                          child: Text("На Вашем счету: ",
+                          padding: EdgeInsets.only(bottom: 6.0, right: 4.0),
+                          child: Text(
+                            "На Вашем счету: ",
                             style: TextStyle(
                               color: Colors.white,
                               fontSize: 16,
@@ -189,39 +179,39 @@ Widget blueZone_1(Map userInfo) {
                         Text(
                           userInfo["extra_account"].toString() + " р.",
                           style: TextStyle(
-                            color: double.parse(userInfo["extra_account"]) < 0 ? Colors.pink : Colors.white,
+                            color: double.parse(userInfo["extra_account"]) < 0
+                                ? Colors.pink
+                                : Colors.white,
                             fontSize: 32,
                             fontWeight: FontWeight.bold,
                           ),
                         ),
                       ],
                     ),
-                    ),
-                    //Text("${ 1 - secToDate(userInfo['packet_secs']/30)}"),
-                  ],
-                ),
-
+                  ),
+                  //Text("${ 1 - secToDate(userInfo['packet_secs']/30)}"),
+                ],
               ),
             ),
-          ],
-        ),
+          ),
+        ],
       ),
-    );
+    ),
+  );
 }
+
 Widget blueZone_2(Map userInfo) {
-  return
-    Container(
+  return Container(
       child: Column(
-        children: <Widget>[
-          double.parse(userInfo["debt"]) > 0
-              ?
-          Container (
-              decoration: new BoxDecoration (
-                  color: Colors.red
-              ),
+    children: <Widget>[
+      double.parse(userInfo["debt"]) > 0
+          ? Container(
+              decoration: new BoxDecoration(color: Colors.red),
               child: new ListTile(
                 title: Text(
-                  "За вами числится задолженость " + userInfo["debt"]?.toString() + " р.",
+                  "За вами числится задолженость " +
+                      userInfo["debt"]?.toString() +
+                      " р.",
                   style: TextStyle(
                     fontWeight: FontWeight.bold,
                     fontSize: 16,
@@ -229,13 +219,10 @@ Widget blueZone_2(Map userInfo) {
                   ),
                 ),
                 //subtitle: Text(userInfo["tarif_name"] + " (" + userInfo["tarif_sum"]?.toString() + "р.)" ),
-              )
-          )
-              :
-          Container(),
-        ],
-      )
-    );
+              ))
+          : Container(),
+    ],
+  ));
 }
 
 Widget blueZoneT(Map userInfo) {
@@ -248,39 +235,34 @@ Widget blueZoneS(Map userInfo) {
 
 Widget blueZoneM() {
   print(userInfo.toString());
-  return
-    Padding(
+  return Padding(
       padding: EdgeInsets.all(16.0),
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.start,
-        children: <Widget>[
-          /*Center(
+      child:
+          Column(mainAxisAlignment: MainAxisAlignment.start, children: <Widget>[
+        /*Center(
             child: Text('Сообщение в службу поддержки:')
           ),*/
-          Center(
-            child: TextField(
-              maxLines: 4,
-              //inputFormatters: [phone],
-              keyboardType: TextInputType.multiline,
-              onChanged: (String text) {sendText = text;},
-              decoration: InputDecoration(
-                  border: OutlineInputBorder(),
-                  labelText: 'Введите текст сообщения'
-              ),
-            ),
+        Center(
+          child: TextField(
+            maxLines: 4,
+            //inputFormatters: [phone],
+            keyboardType: TextInputType.multiline,
+            onChanged: (String text) {
+              sendText = text;
+            },
+            decoration: InputDecoration(
+                border: OutlineInputBorder(),
+                labelText: 'Введите текст сообщения'),
           ),
-          Center(
-            child: RaisedButton(
-              onPressed: sendMessagePressed,
-              child: Text('Отправить в службу поддержки')
-            )
-          )
-        ]
-      )
-    );
+        ),
+        Center(
+            child: ElevatedButton(
+                onPressed: sendMessagePressed,
+                child: Text('Отправить в службу поддержки')))
+      ]));
 }
 
-void sendMessagePressed() async{
+void sendMessagePressed() async {
   String _url = 'https://app.evpanet.com/?set=add_request';
   _url += '&guid=${guids[currentGuidIndex]}';
   _url += '&devid=$devKey';
