@@ -1,22 +1,27 @@
 import 'package:flutter/material.dart';
+import 'package:fluttertoast/fluttertoast.dart';
 import 'package:myevpanet/main.dart';
 import 'package:responsive_flutter/responsive_flutter.dart';
 import 'package:url_launcher/url_launcher.dart';
 import 'package:myevpanet/api/api.dart';
+import 'package:myevpanet/helpers/data.dart';
 
 class CircleButton extends StatelessWidget {
   final GestureTapCallback onTap;
   final IconData iconData;
+  final double size;
 
-  const CircleButton({Key key, this.onTap, this.iconData}) : super(key: key);
+  const CircleButton({Key key, this.onTap, this.iconData, this.size})
+      : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    double size = 35.0;
+    //double size = 35.0;
 
     return new InkResponse(
       onTap: onTap,
       child: new Container(
+        //padding: EdgeInsets.only(top: size - 25.0),
         width: size,
         height: size,
         decoration: new BoxDecoration(
@@ -36,58 +41,55 @@ class CallWindowModal extends StatefulWidget {
   @override
   CallButtonWidget createState() => CallButtonWidget();
 }
+
 class CallButtonWidget extends State {
   String phoneToCall = '+79780489664';
 
   @override
   Widget build(BuildContext context) {
-    return
-      Container(
-        margin: EdgeInsets.only(left: 0.0,right: 0.0),
-        child: Stack(
-          children: <Widget>[
-            Container(
-              padding: EdgeInsets.only(
-                top: 18.0,
-              ),
-              margin: EdgeInsets.only(top: 13.0,right: 8.0),
-              decoration: BoxDecoration(
-                  color: Colors.white,
-                  shape: BoxShape.rectangle,
-                  borderRadius: BorderRadius.circular(6.0),
-                  boxShadow: <BoxShadow>[
-                    BoxShadow(
-                      color: Colors.black54,
-                      blurRadius: 0.0,
-                      offset: Offset(0.0, 0.0),
-                    ),
-                  ]
-              ),
-              child: Column(
-                mainAxisSize: MainAxisSize.min,
-                crossAxisAlignment: CrossAxisAlignment.stretch,
-                children: <Widget>[
-                  Container(
-                      padding: EdgeInsets.all(
-                          16.0
-                      ),
-                      child: Column(
-                        children: <Widget>[
-                          Center(
-                           child: Text(
-                             "Пожалуйста выберите один из номеров технической поддержки.",
-                             style: TextStyle(
-                                 fontSize: ResponsiveFlutter.of(context).fontSize(2),
-                                 color: Color.fromRGBO(72, 95, 113, 1.0),
-                                 fontWeight: FontWeight.bold
-                             ),
-                             textAlign: TextAlign.center,
-                           ),
+    return Container(
+      margin: EdgeInsets.only(left: 0.0, right: 0.0),
+      child: Stack(
+        children: <Widget>[
+          Container(
+            padding: EdgeInsets.only(
+              top: 18.0,
+            ),
+            margin: EdgeInsets.only(top: 13.0, right: 8.0),
+            decoration: BoxDecoration(
+                color: Colors.white,
+                shape: BoxShape.rectangle,
+                borderRadius: BorderRadius.circular(6.0),
+                boxShadow: <BoxShadow>[
+                  BoxShadow(
+                    color: Colors.black54,
+                    blurRadius: 0.0,
+                    offset: Offset(0.0, 0.0),
+                  ),
+                ]),
+            child: Column(
+              mainAxisSize: MainAxisSize.min,
+              crossAxisAlignment: CrossAxisAlignment.stretch,
+              children: <Widget>[
+                Container(
+                    padding: EdgeInsets.all(16.0),
+                    child: Column(
+                      children: <Widget>[
+                        Center(
+                          child: Text(
+                            "Пожалуйста выберите один из номеров технической поддержки.",
+                            style: TextStyle(
+                                fontSize:
+                                    ResponsiveFlutter.of(context).fontSize(2),
+                                color: Color.fromRGBO(72, 95, 113, 1.0),
+                                fontWeight: FontWeight.bold),
+                            textAlign: TextAlign.center,
                           ),
-                          Padding(
-                            padding: EdgeInsets.all(6.0),
-                          ),
-                          DropdownButton<String>(
+                        ),
+                        Padding(
+                          padding: EdgeInsets.all(6.0),
+                        ),
+                        DropdownButton<String>(
                             value: phoneToCall,
                             autofocus: true,
                             icon: Icon(Icons.arrow_drop_down),
@@ -109,9 +111,8 @@ class CallButtonWidget extends State {
                                     ],
                                   )
 
-
                                   //child: Text('+7 (978) 048-96-64')
-                              ),
+                                  ),
                               DropdownMenuItem(
                                   value: '+79780755900',
                                   child: Row(
@@ -120,55 +121,53 @@ class CallButtonWidget extends State {
                                       SizedBox(width: 20.0),
                                       Text('+7 (978) 075-59-00')
                                     ],
-                                  )
-                              ),
-                            ]
-                          ),
-                        ],
-                      )
-                  ),
-                  SizedBox(height: 24.0),
-                  InkWell(
-                    child: Container(
-                      padding: EdgeInsets.only(top: 15.0,bottom:15.0),
-                      decoration: BoxDecoration(
-                        color: Color(0xff374b5d),
-                        borderRadius: BorderRadius.only(
-                            bottomLeft: Radius.circular(6.0),
-                            bottomRight: Radius.circular(6.0)),
-                      ),
-                      child:  Text(
-                        "Совершить звонок",
-                        style: TextStyle(color: Colors.white,fontSize: 20.0),
-                        textAlign: TextAlign.center,
-                      ),
+                                  )),
+                            ]),
+                      ],
+                    )),
+                SizedBox(height: 24.0),
+                InkWell(
+                  child: Container(
+                    padding: EdgeInsets.only(top: 15.0, bottom: 15.0),
+                    decoration: BoxDecoration(
+                      color: Color(0xff374b5d),
+                      borderRadius: BorderRadius.only(
+                          bottomLeft: Radius.circular(6.0),
+                          bottomRight: Radius.circular(6.0)),
                     ),
-                    onTap: () async{
-                      if (await canLaunch('tel://$phoneToCall')) launch('tel://$phoneToCall');
-                    },
-                  )
-                ],
-              ),
-            ),
-            Positioned(
-              right: 0.0,
-              child: GestureDetector(
-                onTap: (){
-                  Navigator.of(context).pop();
-                },
-                child: Align(
-                  alignment: Alignment.topRight,
-                  child: CircleAvatar(
-                    radius: 14.0,
-                    backgroundColor: Colors.red,
-                    child: Icon(Icons.close, color: Colors.white),
+                    child: Text(
+                      "Совершить звонок",
+                      style: TextStyle(color: Colors.white, fontSize: 20.0),
+                      textAlign: TextAlign.center,
+                    ),
                   ),
+                  onTap: () async {
+                    if (await canLaunch('tel://$phoneToCall'))
+                      launch('tel://$phoneToCall');
+                  },
+                )
+              ],
+            ),
+          ),
+          Positioned(
+            right: 0.0,
+            child: GestureDetector(
+              onTap: () {
+                Navigator.of(context).pop();
+              },
+              child: Align(
+                alignment: Alignment.topRight,
+                child: CircleAvatar(
+                  radius: 14.0,
+                  backgroundColor: Colors.red,
+                  child: Icon(Icons.close, color: Colors.white),
                 ),
               ),
             ),
-          ],
-        ),
-      );
+          ),
+        ],
+      ),
+    );
   }
 }
 
@@ -176,20 +175,21 @@ class SupportMessageModal extends StatefulWidget {
   @override
   SupportMessageModalWidget createState() => SupportMessageModalWidget();
 }
+
 class SupportMessageModalWidget extends State {
   String text = '';
 
   @override
   Widget build(BuildContext context) {
     return Container(
-      margin: EdgeInsets.only(left: 0.0,right: 0.0),
+      margin: EdgeInsets.only(left: 0.0, right: 0.0),
       child: Stack(
         children: <Widget>[
           Container(
             padding: EdgeInsets.only(
               top: 18.0,
             ),
-            margin: EdgeInsets.only(top: 13.0,right: 8.0),
+            margin: EdgeInsets.only(top: 13.0, right: 8.0),
             decoration: BoxDecoration(
                 color: Colors.white,
                 shape: BoxShape.rectangle,
@@ -200,18 +200,14 @@ class SupportMessageModalWidget extends State {
                     blurRadius: 0.0,
                     offset: Offset(0.0, 0.0),
                   ),
-                ]
-            ),
+                ]),
             child: Column(
               mainAxisSize: MainAxisSize.min,
               crossAxisAlignment: CrossAxisAlignment.stretch,
               children: <Widget>[
                 Container(
-                  padding: EdgeInsets.only(
-                      left: 16.0,
-                      right: 16.0,
-                      bottom: 20.0
-                  ),
+                  padding:
+                      EdgeInsets.only(left: 16.0, right: 16.0, bottom: 20.0),
                   alignment: Alignment.topCenter,
                   child: Center(
                     child: Text(
@@ -219,8 +215,7 @@ class SupportMessageModalWidget extends State {
                       style: TextStyle(
                           fontSize: ResponsiveFlutter.of(context).fontSize(2),
                           color: Color.fromRGBO(72, 95, 113, 1.0),
-                          fontWeight: FontWeight.bold
-                      ),
+                          fontWeight: FontWeight.bold),
                     ),
                   ),
                 ),
@@ -263,20 +258,20 @@ class SupportMessageModalWidget extends State {
                 SizedBox(height: 24.0),
                 InkWell(
                   child: Container(
-                    padding: EdgeInsets.only(top: 15.0,bottom:15.0),
+                    padding: EdgeInsets.only(top: 15.0, bottom: 15.0),
                     decoration: BoxDecoration(
                       color: Color(0xff374b5d),
                       borderRadius: BorderRadius.only(
                           bottomLeft: Radius.circular(6.0),
                           bottomRight: Radius.circular(6.0)),
                     ),
-                    child:  Text(
+                    child: Text(
                       "Отправить сообщение",
-                      style: TextStyle(color: Colors.white,fontSize: 20.0),
+                      style: TextStyle(color: Colors.white, fontSize: 20.0),
                       textAlign: TextAlign.center,
                     ),
                   ),
-                  onTap:(){
+                  onTap: () {
                     _sendMessagePressed();
                   },
                 )
@@ -286,7 +281,7 @@ class SupportMessageModalWidget extends State {
           Positioned(
             right: 0.0,
             child: GestureDetector(
-              onTap: (){
+              onTap: () {
                 Navigator.of(context).pop();
               },
               child: Align(
@@ -303,10 +298,33 @@ class SupportMessageModalWidget extends State {
       ),
     );
   }
-  void _sendMessagePressed() async{
-    String answer = await RestAPI().remontAddPOST(text, guids[currentGuidIndex], devKey);
-    print(answer);
+
+  void _sendMessagePressed() async {
+    String answer =
+        await RestAPI().remontAddPOST(text, guids[currentGuidIndex], devKey);
+    dprintL(answer);
+    Pushes _pushes = Pushes();
+    _pushes.loadSavedPushes().then((value) {
+      OneNotification _push = OneNotification();
+      int id = int.parse(users[currentGuidIndex]['id']);
+      _push.id = id;
+      _push.title = '($id) Сообщение в службу поддержки';
+      _push.body = text;
+      _push.date = DateTime.now().toLocal().toString();
+      _push.seen = false;
+      _pushes.pushes.add(_push);
+      _pushes.savePushes();
+    });
+
+    Fluttertoast.showToast(
+        msg: "$answer. В случае ответа Вы получите уведомление.",
+        toastLength: Toast.LENGTH_SHORT,
+        gravity: ToastGravity.BOTTOM,
+        timeInSecForIosWeb: 15,
+        webShowClose: true,
+        backgroundColor: Colors.blue,
+        textColor: Colors.white,
+        fontSize: 16.0);
     Navigator.pop(context);
   }
-
 }
