@@ -4,23 +4,20 @@ import 'package:myevpanet/api/api.dart';
 import 'package:myevpanet/main.dart';
 
 class Auth {
-
   bool apiReg(String phone, int id) {
-
     var result = RestAPI().authorizeUserPOST('+' + phone, id, devKey);
     print(result);
     bool _checks = true;
     if (result.toString().contains('00000000-0000-0000-C000-000000000046')) {
       print('Auth failed. No such users.');
       Fluttertoast.showToast(
-      msg: "Такой пользователь не найден",
-      toastLength: Toast.LENGTH_SHORT,
-      gravity: ToastGravity.CENTER,
-      timeInSecForIosWeb: 2,
-      backgroundColor: Colors.red,
-      textColor: Colors.white,
-      fontSize: 16.0
-      );
+          msg: "Такой пользователь не найден",
+          toastLength: Toast.LENGTH_SHORT,
+          gravity: ToastGravity.CENTER,
+          timeInSecForIosWeb: 2,
+          backgroundColor: Colors.red,
+          textColor: Colors.white,
+          fontSize: 16.0);
       _checks = false;
     }
     print(result.toString().length);
