@@ -1,13 +1,12 @@
+//import 'dart:async';
+import 'dart:math';
 import 'dart:ui';
 import 'package:myevpanet/helpers/DesignHelper.dart';
-//import 'package:myevpanet/payment_screen/paymaster.dart';
 import 'package:myevpanet/push_screen/pushList.dart';
 import 'package:myevpanet/webview_screens/pay_widget.dart';
 import 'package:myevpanet/widgets/drawer.dart';
 import 'package:responsive_flutter/responsive_flutter.dart';
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/widgets.dart';
 import 'package:flutter_icons/flutter_icons.dart';
 import 'package:myevpanet/main.dart';
 import 'package:myevpanet/api/api.dart';
@@ -17,9 +16,8 @@ import 'package:intl/intl.dart';
 import 'package:intl/date_symbol_data_local.dart';
 import 'package:myevpanet/main_screen/setups.dart';
 
-
 class MainScreenWidget extends StatefulWidget {
-  
+  //final Pushes _pushes = Pushes();
   MainScreenWidget({Key key, this.title}) : super(key: key);
   final String title;
 
@@ -27,19 +25,45 @@ class MainScreenWidget extends StatefulWidget {
   _MainScreenWidgetState createState() => _MainScreenWidgetState();
 }
 
-class _MainScreenWidgetState extends State<MainScreenWidget> {
-
+class _MainScreenWidgetState extends State<MainScreenWidget>
+    with TickerProviderStateMixin {
   String text = '';
   String phoneToCall = '+79780489664';
   DateFormat dateFormat;
+  //Pushes pushes;
+  bool isAllPushesSeen = true;
+  AnimationController controller;
+  Animation animation;
 
   @override
   void initState() {
-    if (verbose >= 1) print('MainScreen initState()');
-    if (verbose >= 1) print('Start getData');
+    dprintL('MainScreen initState()');
+    dprintL('Start getData');
     getData();
+    Pushes pushes = Pushes();
+    //pushes.loadSavedPushes().then((value) {
+    isAllPushesSeen = pushes.isAllSeen();
+    setState(() {});
+    //});
+    guids.forEach((guid) {
+      pushes
+          .getNewUnreadPushesFromServer(
+              devKey, guid, int.parse(users[currentGuidIndex]['id']))
+          .then((value) {
+        isAllPushesSeen = pushes.isAllSeen();
+        setState(() {});
+      });
+    });
     super.initState();
-    //fbHelper.configure(this.context);
+    controller =
+        AnimationController(duration: const Duration(seconds: 3), vsync: this)
+          ..repeat();
+    animation = Tween(begin: 0.0, end: 1.0).animate(controller)
+      ..addListener(() {
+        setState(() {
+          //isAllPushesSeen = pushes.isAllSeen();
+        });
+      });
     if (verbose >= 1) print('End getData');
     Intl.defaultLocale = 'ru_RU';
     initializeDateFormatting();
@@ -47,10 +71,11 @@ class _MainScreenWidgetState extends State<MainScreenWidget> {
       systemNavigationBarColor: Colors.transparent, // navigation bar color
       statusBarColor: Colors.transparent, // status bar color
       statusBarIconBrightness: Brightness.dark, // status bar icons' color
-      systemNavigationBarIconBrightness: Brightness.dark, //navigation bar icons' color
+      systemNavigationBarIconBrightness:
+          Brightness.dark, //navigation bar icons' color
     ));
-
   }
+
   void getData() async {
     await UserInfo().getUserData();
     setState(() {});
@@ -60,17 +85,23 @@ class _MainScreenWidgetState extends State<MainScreenWidget> {
   int _current = 0;
 
   @override
+  void dispose() {
+    super.dispose();
+    controller.dispose();
+    //contrller2.dispose();
+  }
+
+  @override
   Widget build(BuildContext context) {
     return Scaffold(
-      drawer: AppDrawer(),
-      resizeToAvoidBottomPadding: false,
-      backgroundColor: Color.fromRGBO(245, 246, 248, 1.0),
-      appBar: PreferredSize(
-        preferredSize: Size.fromHeight(70.0), // here the desired height
-        child: AppBar(
-            iconTheme: new IconThemeData(
-                color: Color.fromRGBO(72, 95, 113, 1.0)
-            ),
+        drawer: AppDrawer(),
+        resizeToAvoidBottomInset: false,
+        backgroundColor: Color.fromRGBO(245, 246, 248, 1.0),
+        appBar: PreferredSize(
+          preferredSize: Size.fromHeight(70.0), // here the desired height
+          child: AppBar(
+            iconTheme:
+                new IconThemeData(color: Color.fromRGBO(72, 95, 113, 1.0)),
             titleSpacing: 0.0,
             brightness: Brightness.light,
             backgroundColor: Color.fromRGBO(245, 246, 248, 1.0),
@@ -81,92 +112,90 @@ class _MainScreenWidgetState extends State<MainScreenWidget> {
                 Text(
                   "Информация",
                   style: TextStyle(
-                      color: Color.fromRGBO(72, 95, 113, 1.0),
-                      fontSize: 24.0
-                  ),
+                      color: Color.fromRGBO(72, 95, 113, 1.0), fontSize: 24.0),
                 ),
                 Text(
                   DateFormat.yMMMMd().format(DateTime.now()),
                   style: TextStyle(
                       color: Color.fromRGBO(146, 152, 166, 1.0),
-                      fontSize: 14.0
-                  ),
+                      fontSize: 14.0),
                 )
               ],
-            ),              // Your widgets here
+            ), // Your widgets here
             elevation: 0.0,
             actions: <Widget>[
               GestureDetector(
                 child: Container(
-                    padding: EdgeInsets.only(
-                        top: 10.0,
-                        bottom: 10.0,
-                        right: 10.0
-                    ),
+                    padding:
+                        EdgeInsets.only(top: 10.0, bottom: 10.0, right: 10.0),
                     child: Icon(
                       MaterialCommunityIcons.phone,
                       color: Color.fromRGBO(72, 95, 113, 1.0),
                       size: 24.0,
-                    )
-                ),
+                    )),
                 onTap: () {
                   _showModalCallSupport();
                 },
               ),
               GestureDetector(
                 child: Container(
-                    padding: EdgeInsets.only(
-                      top: 10.0,
-                      bottom: 10.0,
-                      right: 16.0
-                    ),
+                    padding:
+                        EdgeInsets.only(top: 10.0, bottom: 10.0, right: 16.0),
                     child: Icon(
                       MaterialCommunityIcons.face_agent,
                       color: Color.fromRGBO(72, 95, 113, 1.0),
                       size: 24.0,
-                    )
-                ),
+                    )),
                 onTap: () {
                   _showModalSupport();
                 },
               ),
               GestureDetector(
                 child: Container(
-                    padding: EdgeInsets.only(
-                        top: 10.0,
-                        bottom: 10.0,
-                        right: 16.0,
-                        left: 16.0
-                    ),
-                    child: Icon(
-                      MaterialCommunityIcons.bell,
-                      color: Color.fromRGBO(72, 95, 113, 1.0),
-                      size: 24.0,
-                    )
+                  padding: EdgeInsets.only(
+                      top: 10.0, bottom: 10.0, right: 16.0, left: 16.0),
+                  child: isAllPushesSeen
+                      ? Icon(Icons.notifications_none_outlined, size: 24.0)
+                      : Transform(
+                          alignment: FractionalOffset.center,
+                          transform: Matrix4.identity()
+                            //..setEntry(3, 2, 0.002)
+                            ..rotateY(pi * animation.value),
+                          //animation: animation,
+                          child: Icon(Icons.notifications_active, size: 24.0),
+                        ), //MaterialCommunityIcons.bell_ring,
+                  //color: Color.fromRGBO(72, 95, 113, 1.0),
                 ),
-                onTap: () {
-                  Navigator.of(context).push(MaterialPageRoute(builder: (BuildContext context) => PushScreen()));
+                onTap: () async {
+                  //await pushes.loadSavedPushes();
+                  await Navigator.of(context).push(MaterialPageRoute(
+                      builder: (BuildContext context) => PushScreen()));
+                  Pushes pushes = Pushes();
+                  await pushes.loadSavedPushes();
+                  setState(() {
+                    isAllPushesSeen = pushes.isAllSeen();
+                  });
                 },
               ),
             ],
+          ),
         ),
-      ),
-      body: Column(
-        children: [
+        body: Column(children: [
           // каруселька
           Container(
             child: CarouselSlider(
               items: idList(),
+              options: CarouselOptions(
               autoPlay: false,
-              enlargeCenterPage: true,
-              aspectRatio: 16/10,
+              enlargeCenterPage: false,
+              aspectRatio: 16 / 10,
               viewportFraction: 0.85,
-              onPageChanged: (index) {
+              onPageChanged: (index, reason) {
                 currentGuidIndex = index;
                 userInfo = users[currentGuidIndex];
                 _current = index;
                 setState(() {});
-              },
+              }),
             ),
           ),
           // навигационные точечки
@@ -174,147 +203,127 @@ class _MainScreenWidgetState extends State<MainScreenWidget> {
             //padding: EdgeInsets.all(5.0),
             child: Row(
                 mainAxisAlignment: MainAxisAlignment.center,
-                children: idListPoints()
-            ),
+                children: idListPoints()),
           ),
           // секция с картами деталей учетной записи
           Expanded(
             child: ListView(
               physics: ScrollPhysics(parent: BouncingScrollPhysics()),
-              padding: EdgeInsets.only(
-                top: 0.0,
-                left: 30.0,
-                right: 30.0
-              ),
-              children: <Widget> [
+              padding: EdgeInsets.only(top: 0.0, left: 30.0, right: 30.0),
+              children: <Widget>[
                 // виджет отображения долга
                 Container(
                     child: Column(
-                      children: <Widget>[
-                        double.parse(users[currentGuidIndex]["debt"]) > 0
-                            ?
-                            Card(
-                                color: Colors.red,
-                                child: new ListTile(
-                                      title: Text(
-                                        "За Вашей учётной записью числится задолженость ${users[currentGuidIndex]["debt"].toString()} р.",
-                                        style: TextStyle(
-                                          fontWeight: FontWeight.bold,
-                                          fontSize: 16,
-                                          color: Colors.white,
-                                        ),
-                                      ),
-                                )
-                            )
-                            :
-                        Container(),
-                      ],
-                    )
-                ),
+                  children: <Widget>[
+                    double.parse(users[currentGuidIndex]["debt"]) > 0
+                        ? Card(
+                            color: Colors.red,
+                            child: new ListTile(
+                              title: Text(
+                                "За Вашей учётной записью числится задолженость ${users[currentGuidIndex]["debt"].toString()} р.",
+                                style: TextStyle(
+                                  fontWeight: FontWeight.bold,
+                                  fontSize: 16,
+                                  color: Colors.white,
+                                ),
+                              ),
+                            ))
+                        : Container(),
+                  ],
+                )),
                 // текст - Детали учетной записи
                 Container(
-                  padding: EdgeInsets.only(
-                    left: 40.0,
-                    top: 10.0,
-                    bottom: 10.0
-                  ),
+                  padding: EdgeInsets.only(left: 40.0, top: 10.0, bottom: 10.0),
                   alignment: Alignment.centerLeft,
                   child: Text(
                     'Детали учётной записи',
                     style: TextStyle(
-                      fontSize: ResponsiveFlutter.of(context).fontSize(2.7),
-                      color: Color.fromRGBO(72, 95, 113, 1.0),
-                      fontWeight: FontWeight.bold
+                        fontSize: ResponsiveFlutter.of(context).fontSize(2.7),
+                        color: Color.fromRGBO(72, 95, 113, 1.0),
+                        fontWeight: FontWeight.bold),
+                  ),
+                ),
+                Container(
+                  padding: EdgeInsets.only(top: 5.0),
+                  child: Card(
+                      child: Container(
+                    child: Column(
+                      mainAxisSize: MainAxisSize.min,
+                      children: <Widget>[
+                        ListTile(
+                          leading: Icon(Icons.album, size: 40),
+                          title: Text('Тарифный план'),
+                          subtitle: Text(
+                              '${users[currentGuidIndex]["tarif_name"]} (${users[currentGuidIndex]["tarif_sum"].toString()} р.)'),
+                        ),
+                      ],
                     ),
-                  ),
+                  )),
                 ),
                 Container(
-                  padding: EdgeInsets.only(
-                      top: 5.0),
+                  padding: EdgeInsets.only(top: 5.0),
                   child: Card(
                       child: Container(
-                        child: Column(
-                          mainAxisSize: MainAxisSize.min,
-                          children: <Widget>[
-                            ListTile(
-                              leading: Icon(Icons.album, size: 40),
-                              title: Text('Тарифный план'),
-                              subtitle: Text(
-                                  '${users[currentGuidIndex]["tarif_name"]} (${users[currentGuidIndex]["tarif_sum"].toString()} р.)'
-                              ),
-                            ),
-                          ],
+                    child: Column(
+                      mainAxisSize: MainAxisSize.min,
+                      children: <Widget>[
+                        ListTile(
+                          leading: Icon(Icons.album, size: 40),
+                          title: Text('IP адрес абонента'),
+                          subtitle: Text(users[currentGuidIndex]["real_ip"]),
                         ),
-                      )
-                  ),
+                      ],
+                    ),
+                  )),
                 ),
                 Container(
-                  padding: EdgeInsets.only(
-                      top: 5.0),
+                  padding: EdgeInsets.only(top: 5.0),
                   child: Card(
                       child: Container(
-                        child: Column(
-                          mainAxisSize: MainAxisSize.min,
-                          children: <Widget>[
-                            ListTile(
-                              leading: Icon(Icons.album, size: 40),
-                              title: Text('IP адрес абонента'),
-                              subtitle: Text(users[currentGuidIndex]["real_ip"]),
-                            ),
-                          ],
+                    child: Column(
+                      mainAxisSize: MainAxisSize.min,
+                      children: <Widget>[
+                        ListTile(
+                          leading: Icon(Icons.album, size: 40),
+                          title: Text('Адрес подключения'),
+                          subtitle: Text(users[currentGuidIndex]["street"] +
+                              ", д. " +
+                              users[currentGuidIndex]["house"] +
+                              ", кв. " +
+                              users[currentGuidIndex]["flat"]),
                         ),
-                      )
-                  ),
-                ),
-                Container(
-                  padding: EdgeInsets.only(
-                      top: 5.0),
-                  child: Card(
-                      child: Container(
-                        child: Column(
-                          mainAxisSize: MainAxisSize.min,
-                          children: <Widget>[
-                            ListTile(
-                              leading: Icon(Icons.album, size: 40),
-                              title: Text('Адрес подключения'),
-                              subtitle: Text(users[currentGuidIndex]["street"] + ", д. " + users[currentGuidIndex]["house"] + ", кв. " + users[currentGuidIndex]["flat"]),
-                            ),
-                          ],
-                        ),
-                      )
-                  ),
+                      ],
+                    ),
+                  )),
                 ),
               ],
             ),
           )
-        ]
-      )
-    );
+        ]));
   }
+
   // генерируем точки
   List<Widget> idListPoints() {
     List<Widget> _list = [];
-    for (var item = 0; item < guids.length; item++/*users.keys*/) {
-      _list.add(
-        Container(
-          width: 8.0,
-          height: 8.0,
-          margin: EdgeInsets.symmetric(vertical: 10.0, horizontal: 2.0),
-          decoration: BoxDecoration(
-              shape: BoxShape.circle,
-              color: _current == item
-                  ? Color.fromRGBO(116, 162, 177, 1.0)
-                  : Color.fromRGBO(198, 209, 216, 1.0)
-          ),
-        )
-      );
+    for (var item = 0; item < guids.length; item++ /*users.keys*/) {
+      _list.add(Container(
+        width: 8.0,
+        height: 8.0,
+        margin: EdgeInsets.symmetric(vertical: 10.0, horizontal: 2.0),
+        decoration: BoxDecoration(
+            shape: BoxShape.circle,
+            color: _current == item
+                ? Color.fromRGBO(116, 162, 177, 1.0)
+                : Color.fromRGBO(198, 209, 216, 1.0)),
+      ));
     }
     return _list;
   }
+
   // список тёмных плашек
   List<Widget> idList() {
     List<Widget> _list = [];
-    for (var item = 0; item < guids.length; item++/*users.keys*/) {
+    for (var item = 0; item < guids.length; item++ /*users.keys*/) {
       var now = DateTime.now().toUtc();
       var packet = DateTime.parse(users[item]["packet_end_utc"]).toUtc();
       var daysRemain = packet.difference(now).inDays;
@@ -322,30 +331,27 @@ class _MainScreenWidgetState extends State<MainScreenWidget> {
       var packetEndColor = Colors.white;
 
       if (daysRemain > 0 && daysRemain < 1) {
-       packetEndColor = Colors.amber;
-      }
-      else if (daysRemain <= 0) {
-       packetEndColor = Colors.red;
+        packetEndColor = Colors.amber;
+      } else if (daysRemain <= 0) {
+        packetEndColor = Colors.red;
       }
 
-        _list.add(
-          GestureDetector(
-            onTap: () async{
-              print('!!!!!!!!! current index is: $currentGuidIndex');
-              await Navigator.of(context).push(MaterialPageRoute(builder: (BuildContext context) => SetupGroup()));
-              setState(() {});
-            },
-            child: Container(
-              padding: EdgeInsets.only(
-                  top: 0.0,
-                  left: 5.0,
-                  right: 5.0,
-                  bottom: 0.0
-              ),
-                child: Container(
-                  decoration: BoxDecoration(
-                    border: Border.all(width: 1.0, color: Color.fromRGBO(52, 79, 100, 1.0)),
-                    borderRadius: BorderRadius.circular(ResponsiveFlutter.of(context).moderateScale(8)),
+      _list.add(GestureDetector(
+          onTap: () async {
+            //print('!!!!!!!!! current index is: $currentGuidIndex');
+            await Navigator.of(context).push(MaterialPageRoute(
+                builder: (BuildContext context) => SetupGroup()));
+            setState(() {});
+          },
+          child: Container(
+              padding:
+                  EdgeInsets.only(top: 0.0, left: 5.0, right: 5.0, bottom: 0.0),
+              child: Container(
+                decoration: BoxDecoration(
+                    border: Border.all(
+                        width: 1.0, color: Color.fromRGBO(52, 79, 100, 1.0)),
+                    borderRadius: BorderRadius.circular(
+                        ResponsiveFlutter.of(context).moderateScale(8)),
                     boxShadow: [
                       BoxShadow(
                         color: Color.fromRGBO(184, 202, 220, 1.0),
@@ -364,81 +370,35 @@ class _MainScreenWidgetState extends State<MainScreenWidget> {
                           0.2,
                           1.0,
                         ],
-                        colors: [Color.fromRGBO(68, 98, 124, 1), Color.fromRGBO(10, 33, 51, 1)]
-                    )
-                  ),
-                  child: Container(
-                    padding: EdgeInsets.all(ResponsiveFlutter.of(context).moderateScale(20)),
-                    child: Column(
-                      crossAxisAlignment: CrossAxisAlignment.stretch,
-                      children: <Widget>[
-                        Expanded(
-                          flex: 3,
-                          child: Row(
-                            children: <Widget>[
-                              Expanded(
-                                // Левая сторона верхнего блока
-                                flex: 2,
-                                child: Column(
-                                  crossAxisAlignment: CrossAxisAlignment.start,
-                                  mainAxisAlignment: MainAxisAlignment.start,
-                                  children: <Widget>[
-                                    Container(
-                                      padding: EdgeInsets.all(5.0),
-                                      child: Text(
-                                        'ID: ' + users[item]['id'].toString(),
-                                        style: TextStyle(
-                                          color: Colors.white,
-                                          fontSize: ResponsiveFlutter.of(context).fontSize(2.4),
-                                          fontWeight: FontWeight.bold,
-                                          shadows: [
-                                            Shadow(
-                                              blurRadius: 1.0,
-                                              color: Colors.black,
-                                              offset: Offset(1.0, 1.0),
-                                            ),
-                                          ],
-                                        ),
-                                      ),
-                                    ),
-                                    _current == _current/*item*/ ?
-                                    CircleButton(
-                                      onTap: () {
-                                        Navigator.of(context).push(MaterialPageRoute(builder: (BuildContext context) => PayView()));
-                                      },
-                                      iconData: MaterialCommunityIcons.wallet_plus_outline
-                                    ): Text(''),
-                                  ],
-                                ),
-                              ),
-                              Expanded(
-                                // Правая сторона верхнего блока
-                                flex: 3,
-                                child: Column(
-                                  crossAxisAlignment: CrossAxisAlignment.end,
-                                  mainAxisAlignment: MainAxisAlignment.start,
-                                  children: <Widget>[
-                                    Padding(
-                                      padding: EdgeInsets.only(top: 2.0),
-                                    ),
-                                    Padding(
-                                      padding: EdgeInsets.only(
-                                          bottom: 3.0,
-                                          right: 4.0
-                                      ),
-                                      child: Text("Доступный баланс",
-                                        style: TextStyle(
-                                          color: Color.fromRGBO(144, 198, 124, 1),
-                                          fontSize: ResponsiveFlutter.of(context).fontSize(1.6),
-                                        ),
-                                      ),
-                                    ),
-                                    Text(
-                                      NumberFormat('#,##0.00##', 'ru_RU').format(double.parse(users[item]["extra_account"])) + " р.",
-                                      //userInfo["extra_account"],
+                        colors: [
+                          Color.fromRGBO(68, 98, 124, 1),
+                          Color.fromRGBO(10, 33, 51, 1)
+                        ])),
+                child: Container(
+                  padding: EdgeInsets.all(
+                      ResponsiveFlutter.of(context).moderateScale(20)),
+                  child: Column(
+                    crossAxisAlignment: CrossAxisAlignment.stretch,
+                    children: <Widget>[
+                      Expanded(
+                        flex: 3,
+                        child: Row(
+                          children: <Widget>[
+                            Expanded(
+                              // Левая сторона верхнего блока
+                              flex: 2,
+                              child: Column(
+                                crossAxisAlignment: CrossAxisAlignment.start,
+                                mainAxisAlignment: MainAxisAlignment.start,
+                                children: <Widget>[
+                                  Container(
+                                    padding: EdgeInsets.all(5.0),
+                                    child: Text(
+                                      'ID: ' + users[item]['id'].toString(),
                                       style: TextStyle(
-                                        color: double.parse(users[item]["extra_account"]) < 0 ? Color.fromRGBO(255, 81, 105, 1) : Colors.white,
-                                        fontSize: ResponsiveFlutter.of(context).fontSize(3.2),
+                                        color: Colors.white,
+                                        fontSize: ResponsiveFlutter.of(context)
+                                            .fontSize(2.4),
                                         fontWeight: FontWeight.bold,
                                         shadows: [
                                           Shadow(
@@ -449,29 +409,58 @@ class _MainScreenWidgetState extends State<MainScreenWidget> {
                                         ],
                                       ),
                                     ),
-                                  ],
-                                ),
-                              )
-                            ],
-                          ),
-                        ),
-                        Expanded(
-                          flex: 3,
-                          child: Column(
-                            crossAxisAlignment: CrossAxisAlignment.stretch,
-                            mainAxisAlignment: MainAxisAlignment.start,
-                            children: <Widget>[
-                              Container(
-                                padding: EdgeInsets.only(
-                                  top: 0.0
-                                ),
-                                  child: Text(
-                                    //textScaleFactor1.toString(),
-                                    '${users[item]['name']}',
-                                    textAlign: TextAlign.right,
+                                  ),
+                                  _current == currentGuidIndex /*item*/ ? CircleButton(
+                                          size: 35.0, //animationSize.value,
+                                          onTap: () {
+                                            Navigator.of(context).push(
+                                                MaterialPageRoute(
+                                                    builder: (BuildContext
+                                                            context) =>
+                                                        PayView()));
+                                          },
+                                          iconData: MaterialCommunityIcons.wallet_plus_outline)
+                                      : Text(''),
+                                ],
+                              ),
+                            ),
+                            Expanded(
+                              // Правая сторона верхнего блока
+                              flex: 3,
+                              child: Column(
+                                crossAxisAlignment: CrossAxisAlignment.end,
+                                mainAxisAlignment: MainAxisAlignment.start,
+                                children: <Widget>[
+                                  Padding(
+                                    padding: EdgeInsets.only(top: 2.0),
+                                  ),
+                                  Padding(
+                                    padding: EdgeInsets.only(
+                                        bottom: 3.0, right: 4.0),
+                                    child: Text(
+                                      "Доступный баланс",
+                                      style: TextStyle(
+                                        color: Color.fromRGBO(144, 198, 124, 1),
+                                        fontSize: ResponsiveFlutter.of(context)
+                                            .fontSize(1.6),
+                                      ),
+                                    ),
+                                  ),
+                                  Text(
+                                    NumberFormat('#,##0.00##', 'ru_RU').format(
+                                            double.parse(
+                                                users[item]["extra_account"])) +
+                                        " р.",
+                                    //userInfo["extra_account"],
                                     style: TextStyle(
-                                      fontSize: ResponsiveFlutter.of(context).fontSize(3),
-                                      color: Color.fromRGBO(166, 187, 204, 1),
+                                      color: double.parse(users[item]
+                                                  ["extra_account"]) <
+                                              0
+                                          ? Color.fromRGBO(255, 81, 105, 1)
+                                          : Colors.white,
+                                      fontSize: ResponsiveFlutter.of(context)
+                                          .fontSize(3.2),
+                                      fontWeight: FontWeight.bold,
                                       shadows: [
                                         Shadow(
                                           blurRadius: 1.0,
@@ -481,62 +470,95 @@ class _MainScreenWidgetState extends State<MainScreenWidget> {
                                       ],
                                     ),
                                   ),
-                              )
-                            ],
-                          ),
+                                ],
+                              ),
+                            )
+                          ],
                         ),
-                        Expanded(
-                          flex: 2,
-                          child: Row(
-                            children: <Widget>[
-                              Expanded(
+                      ),
+                      Expanded(
+                        flex: 3,
+                        child: Column(
+                          crossAxisAlignment: CrossAxisAlignment.stretch,
+                          mainAxisAlignment: MainAxisAlignment.start,
+                          children: <Widget>[
+                            Container(
+                              padding: EdgeInsets.only(top: 0.0),
+                              child: Text(
+                                //textScaleFactor1.toString(),
+                                '${users[item]['name']}',
+                                textAlign: TextAlign.right,
+                                style: TextStyle(
+                                  fontSize:
+                                      ResponsiveFlutter.of(context).fontSize(3),
+                                  color: Color.fromRGBO(166, 187, 204, 1),
+                                  shadows: [
+                                    Shadow(
+                                      blurRadius: 1.0,
+                                      color: Colors.black,
+                                      offset: Offset(1.0, 1.0),
+                                    ),
+                                  ],
+                                ),
+                              ),
+                            )
+                          ],
+                        ),
+                      ),
+                      Expanded(
+                        flex: 2,
+                        child: Row(
+                          children: <Widget>[
+                            Expanded(
                                 flex: 3,
                                 child: Column(
                                   crossAxisAlignment: CrossAxisAlignment.start,
                                   mainAxisAlignment: MainAxisAlignment.end,
                                   children: <Widget>[
                                     Text(
-                                        'Окончание действия пакета',
-                                        style: TextStyle(
-                                          fontSize: ResponsiveFlutter.of(context).fontSize(1.4),
-                                          color: Colors.white,
-                                          shadows: [
-                                            Shadow(
-                                              blurRadius: 1.0,
-                                              color: Colors.black,
-                                              offset: Offset(1.0, 1.0),
-                                            ),
-                                          ],
-                                        ),
+                                      'Окончание действия пакета',
+                                      style: TextStyle(
+                                        fontSize: ResponsiveFlutter.of(context)
+                                            .fontSize(1.4),
+                                        color: Colors.white,
+                                        shadows: [
+                                          Shadow(
+                                            blurRadius: 1.0,
+                                            color: Colors.black,
+                                            offset: Offset(1.0, 1.0),
+                                          ),
+                                        ],
+                                      ),
                                     ),
                                     Text(
-                                      userInfo["packet_end"] + " (" + daysRemain.toString() + " дн.)",
-                                        style: TextStyle(
-                                          fontSize: ResponsiveFlutter.of(context).fontSize(1.8),
-                                          fontWeight: FontWeight.bold,
-                                          color: packetEndColor,
-                                          shadows: [
-                                            Shadow(
-                                              blurRadius: 1.0,
-                                              color: Colors.black,
-                                              offset: Offset(1.0, 1.0),
-                                            ),
-                                          ],
-                                        ),
+                                      userInfo["packet_end"] +
+                                          " (" +
+                                          daysRemain.toString() +
+                                          " дн.)",
+                                      style: TextStyle(
+                                        fontSize: ResponsiveFlutter.of(context)
+                                            .fontSize(1.8),
+                                        fontWeight: FontWeight.bold,
+                                        color: packetEndColor,
+                                        shadows: [
+                                          Shadow(
+                                            blurRadius: 1.0,
+                                            color: Colors.black,
+                                            offset: Offset(1.0, 1.0),
+                                          ),
+                                        ],
+                                      ),
                                     ),
                                   ],
-                                )
-                              ),
-                              Expanded(
+                                )),
+                            Expanded(
                                 flex: 1,
                                 child: Row(
                                   crossAxisAlignment: CrossAxisAlignment.center,
                                   mainAxisAlignment: MainAxisAlignment.end,
                                   children: <Widget>[
                                     Padding(
-                                      padding: EdgeInsets.only(
-                                          top: 20.0
-                                      ),
+                                      padding: EdgeInsets.only(top: 20.0),
                                       child: Icon(
                                         MaterialCommunityIcons.cogs,
                                         color: Colors.white,
@@ -544,57 +566,48 @@ class _MainScreenWidgetState extends State<MainScreenWidget> {
                                       ),
                                     )
                                   ],
-                                )
-                              ),
-                            ],
-                          ),
-                        )
-                      ],
-                    ),
+                                )),
+                          ],
+                        ),
+                      )
+                    ],
                   ),
-                )
-            )
-          )
-        );
+                ),
+              ))));
     }
     return _list;
   }
 
   // Вызов модального окна с сообщением в ремонты
-  void _showModalSupport() async{
+  _showModalSupport() async {
     return showGeneralDialog(
         context: context,
         barrierDismissible: true,
-        barrierLabel: MaterialLocalizations.of(context).modalBarrierDismissLabel,
+        barrierLabel:
+            MaterialLocalizations.of(context).modalBarrierDismissLabel,
         barrierColor: Color(0xff2c4860),
         transitionDuration: const Duration(milliseconds: 200),
-        pageBuilder: (BuildContext buildContext, Animation animation, Animation secondaryAnimation) {
+        pageBuilder: (BuildContext buildContext, Animation animation,
+            Animation secondaryAnimation) {
           return Dialog(
-            backgroundColor: Colors.transparent,
-            child: SupportMessageModal()
-          );
-        }
-    );
+              backgroundColor: Colors.transparent,
+              child: SupportMessageModal());
+        });
   }
+
   //  Вызов модального окна для совершения звонка
-  void _showModalCallSupport() async{
+  _showModalCallSupport() async {
     return showGeneralDialog(
         context: context,
         barrierDismissible: true,
-        barrierLabel: MaterialLocalizations.of(context).modalBarrierDismissLabel,
+        barrierLabel:
+            MaterialLocalizations.of(context).modalBarrierDismissLabel,
         barrierColor: Color(0xff2c4860),
         transitionDuration: const Duration(milliseconds: 200),
-        pageBuilder: (BuildContext buildContext, Animation animation, Animation secondaryAnimation) {
+        pageBuilder: (BuildContext buildContext, Animation animation,
+            Animation secondaryAnimation) {
           return Dialog(
-            backgroundColor: Colors.transparent,
-            child: CallWindowModal()
-          );
-        }
-    );
+              backgroundColor: Colors.transparent, child: CallWindowModal());
+        });
   }
-
 }
-
-
-
-
