@@ -1,105 +1,110 @@
 import 'package:flutter/material.dart';
-//import 'package:flutter/services.dart';
-import '../main.dart';
-//import 'package:shared_preferences/shared_preferences.dart';
+import 'package:myevpanet/api/api.dart';
+import 'package:myevpanet/main.dart';
+import 'package:myevpanet/push_screen/notif.dart';
 
 class PushScreen extends StatefulWidget {
-
-  PushScreen({Key key, this.title}) : super(key: key);
-  final String title;
-
   @override
   _PushScreenState createState() => _PushScreenState();
 }
 
 class _PushScreenState extends State<PushScreen> {
+  Pushes pushes = Pushes();
 
   @override
   void initState() {
+    pushes.loadSavedPushes().then((pushesList) {
+      pushesList.sort((a, b) => b.date.compareTo(a.date));
+      setState(() {});
+    });
+    //pushes = Pushes();
+    //setState(() {});
     super.initState();
   }
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(
-        backgroundColor: Color(0xff32506a),
-        title: Text('Мои уведомления')
-      ),
-      body: ListView(
-        children: listRows()
-      ),
-      /*bottomSheet: FlatButton(
-        onPressed: () async{
-          Clipboard.setData(
-            ClipboardData(
-              text: devKey.toString()
-            )
-          );
-          //final shared = await SharedPreferences.getInstance();
-          //sharedPushes.add(devKey.toString());
-          //shared.setStringList('pushes', sharedPushes);
-        },
-        child: Text('Скопировать ключ в память'),
-      )*/
+      appBar: AppBar(actions: [
+        pushes.pushes.length > 0
+            ? IconButton(
+                icon: Icon(Icons.delete_forever_rounded),
+                onPressed: () async {
+                  pushes.pushes.clear();
+                  await pushes.savePushes();
+                  setState(() {});
+                },
+              )
+            : Text('')
+      ], backgroundColor: Color(0xff32506a), title: Text('Мои уведомления')),
+      body: ListView(children: listRows()),
     );
   }
 
   List<Widget> listRows() {
     List<Widget> _list = [];
-    //print('Reading content of pushes.dat');
-    /*for (var item in sharedPushes) {
-      _list.add(Text(item));
-    }*/
-    for (var _push in pushes) {
-      //print(_push);
+    dprintD('forming pushes widgets list...', verbose);
+    pushes.pushes.forEach((_push) {
+      dprintD(_push, verbose);
       _list.add(
-        Padding(
-          padding: EdgeInsets.only(
-            left: 16.0,
-            right: 16.0
-          ),
-          child: CustomListItemTwo(
-            thumbnail: Container(
-              decoration: const BoxDecoration(
-                  color: Color(0xff42617e),
-                  shape: BoxShape.circle
-              ),
-              alignment: Alignment(0.0, 0.0),
-              child: Icon(
+        GestureDetector(
+          onTap: () async {
+            setState(() {
+              _push.seen = true;
+              pushes.savePushes();
+            });
+            dprintL('go to one push show screen');
+            var toDelete = await Navigator.of(context).push(MaterialPageRoute(
+                builder: (BuildContext context) => OneNotif(_push)));
+            if (toDelete.runtimeType.toString() != 'bool') toDelete = false;
+            if (toDelete) {
+              pushes.pushes.remove(_push);
+              pushes.savePushes();
+              setState(() {});
+            }
+          },
+          child: Padding(
+            padding: EdgeInsets.only(left: 16.0, right: 16.0),
+            child: CustomListItemTwo(
+              thumbnail: Container(
+                decoration: BoxDecoration(
+                    color: _push.seen ? Color(0xff42617e) : Colors.redAccent,
+                    shape: BoxShape.circle),
+                alignment: Alignment(0.0, 0.0),
+                child: Icon(
                   Icons.info_outline,
-                color: Colors.white,
-                size: 32.0,
+                  color: Colors.white,
+                  size: 32.0,
+                ),
               ),
+              title: "ID: " + _push.id.toString(),
+              subtitle: _push.body,
+              author: _push.title,
+              publishDate: _push.date
+                  .toString()
+                  .substring(0, _push.date.toString().length - 7),
+              isFullSubtitle: false,
             ),
-            title: "ID: " + _push['id'].toString(),
-            subtitle: _push['body'],
-            author: _push['title'],
-            publishDate: _push['date'].toString(),
-            //readDuration: '12 mins',
           ),
         ),
-
-
-
       );
       _list.add(Divider());
-    }
+    });
     return _list;
   }
-
 }
 
 class CustomListItemTwo extends StatelessWidget {
-  CustomListItemTwo({
-    Key key,
-    this.thumbnail,
-    this.title,
-    this.subtitle,
-    this.author,
-    this.publishDate,
-    this.readDuration,
-  }) : super(key: key);
+  CustomListItemTwo(
+      {Key key,
+      this.thumbnail,
+      this.title,
+      this.subtitle,
+      this.author,
+      this.publishDate,
+      this.readDuration,
+      this.isFullSubtitle})
+      : super(key: key);
 
   final Widget thumbnail;
   final String title;
@@ -107,18 +112,19 @@ class CustomListItemTwo extends StatelessWidget {
   final String author;
   final String publishDate;
   final String readDuration;
+  final bool isFullSubtitle;
 
   @override
   Widget build(BuildContext context) {
     return Padding(
       padding: const EdgeInsets.symmetric(vertical: 10.0),
       child: SizedBox(
-        height: 90,
+        height: isFullSubtitle ? 190 : 90,
         child: Row(
           crossAxisAlignment: CrossAxisAlignment.start,
           children: <Widget>[
             AspectRatio(
-              aspectRatio: 0.6,
+              aspectRatio: 0.3,
               child: thumbnail,
             ),
             Expanded(
@@ -129,7 +135,7 @@ class CustomListItemTwo extends StatelessWidget {
                   subtitle: subtitle,
                   author: author,
                   publishDate: publishDate,
-                  //readDuration: readDuration,
+                  isFullSubtitle: isFullSubtitle,
                 ),
               ),
             )
@@ -141,20 +147,22 @@ class CustomListItemTwo extends StatelessWidget {
 }
 
 class _ArticleDescription extends StatelessWidget {
-  _ArticleDescription({
-    Key key,
-    this.title,
-    this.subtitle,
-    this.author,
-    this.publishDate,
-    this.readDuration,
-  }) : super(key: key);
+  _ArticleDescription(
+      {Key key,
+      this.title,
+      this.subtitle,
+      this.author,
+      this.publishDate,
+      this.readDuration,
+      this.isFullSubtitle})
+      : super(key: key);
 
   final String title;
   final String subtitle;
   final String author;
   final String publishDate;
   final String readDuration;
+  final bool isFullSubtitle;
 
   @override
   Widget build(BuildContext context) {
@@ -171,14 +179,12 @@ class _ArticleDescription extends StatelessWidget {
                 maxLines: 2,
                 overflow: TextOverflow.ellipsis,
                 style: const TextStyle(
-                  fontWeight: FontWeight.bold,
-                  fontSize: 16.0
-                ),
+                    fontWeight: FontWeight.bold, fontSize: 16.0),
               ),
               const Padding(padding: EdgeInsets.only(bottom: 2.0)),
               Text(
                 '$subtitle',
-                maxLines: 2,
+                maxLines: isFullSubtitle ? null : 2,
                 overflow: TextOverflow.ellipsis,
                 style: const TextStyle(
                   fontSize: 13.0,
