@@ -16,21 +16,24 @@ class SplashWidget extends StatefulWidget {
 }
 
 class _SplashWidgetState extends State<SplashWidget> {
-
   @override
   initState() {
-    if (verbose >=1) print('SplashScreen initState');
     super.initState();
-    //fbHelper = FirebaseHelper();
     whereToGo();
     SystemChrome.setPreferredOrientations([
       DeviceOrientation.portraitUp,
       DeviceOrientation.portraitDown,
     ]);
   }
+
   goGo(int index) async {
-    Timer(Duration(seconds: 2), await Navigator.of(context).pushReplacement(MaterialPageRoute(builder: (BuildContext context) => index == 1 ? MainScreenWidget() : LoginWidget())));
+    Timer(
+        Duration(seconds: 2),
+        await Navigator.of(context).pushReplacement(MaterialPageRoute(
+            builder: (BuildContext context) =>
+                index == 1 ? MainScreenWidget() : LoginWidget())));
   }
+
   int guidsNumber = guids.isEmpty ? 1 : guids.length;
   @override
   Widget build(BuildContext context) {
@@ -43,12 +46,7 @@ class _SplashWidgetState extends State<SplashWidget> {
                 gradient: LinearGradient(
                     begin: Alignment.topLeft,
                     end: Alignment.bottomRight,
-                    colors: [
-                      Color(0xff11273c),
-                      Color(0xff3c5d7c)
-                    ]
-                )
-            ),
+                    colors: [Color(0xff11273c), Color(0xff3c5d7c)])),
           ),
           Column(
             mainAxisAlignment: MainAxisAlignment.start,
@@ -61,11 +59,10 @@ class _SplashWidgetState extends State<SplashWidget> {
                       Expanded(
                         flex: 1,
                         child: Container(
-                          child: Padding(
-                            padding: EdgeInsets.all(24.0),
-                              child: Image.asset("assets/images/splash_logo.png"),
-                          )
-                        ),
+                            child: Padding(
+                          padding: EdgeInsets.all(24.0),
+                          child: Image.asset("assets/images/splash_logo.png"),
+                        )),
                       )
                     ],
                   ),
@@ -78,16 +75,12 @@ class _SplashWidgetState extends State<SplashWidget> {
                   children: <Widget>[
                     Padding(
                       padding: EdgeInsets.only(
-                          top: 20.0,
-                          left: 24.0,
-                        right: 24.0,
-                        bottom: 16.0
-                      ),
+                          top: 20.0, left: 24.0, right: 24.0, bottom: 16.0),
                       child: LinearProgressIndicator(
                           value: currentGuidIndex / guidsNumber,
                           backgroundColor: Color(0xff3c5d7c),
-                          valueColor: AlwaysStoppedAnimation<Color>(Colors.white)
-                      ),
+                          valueColor:
+                              AlwaysStoppedAnimation<Color>(Colors.white)),
                     ),
                     Text(
                       "Загрузка...",
@@ -96,10 +89,8 @@ class _SplashWidgetState extends State<SplashWidget> {
                       style: TextStyle(
                           fontWeight: FontWeight.bold,
                           fontSize: 16.0,
-                          color: Colors.white
-                      ),
+                          color: Colors.white),
                     ),
-
                   ],
                 ),
               )
@@ -113,19 +104,21 @@ class _SplashWidgetState extends State<SplashWidget> {
   void whereToGo() async {
     final file = await FileStorage('key.dat').localFile;
     if (file.existsSync()) {
-      if (verbose >=1) print('Key file is exists');
+      dprintL('Key file is exists');
       devKey = file.readAsStringSync(encoding: utf8);
       if (devKey.toString().length > 7) {
-        if (verbose >=1) print('Device key from file is: $devKey');
+        dprintL('Device key from file is: $devKey');
         //пробуем обновить токен гугла
+
         FirebaseHelper().getAppToken().then((String value) {
           if (value == null) {
-            print('we got null token from google');
+            dprintL('we got null token from google');
           } else {
-            print('new token from google is: $value');
+            dprintL('new token from google is: $value');
             if (value != devKey) {
-              print('token has changed. we have to go to new registration');
-              file.writeAsStringSync(value, mode: FileMode.write, encoding: utf8);
+              dprintL('token has changed. we have to go to new registration');
+              file.writeAsStringSync(value,
+                  mode: FileMode.write, encoding: utf8);
               goGo(0);
             }
           }
@@ -134,25 +127,28 @@ class _SplashWidgetState extends State<SplashWidget> {
         //загрузка списка ГУИДОВ из файла
         final _guidsfile = await FileStorage('guidlist.dat').localFile;
         if (_guidsfile.existsSync()) {
-          if (verbose >=1) print('GUIDs file exists');
+          if (verbose >= 1) print('GUIDs file exists');
           var _readResult = _guidsfile.readAsStringSync(encoding: utf8);
           final parsed = jsonDecode(_readResult);
           //json.decode(_readResult);
-          if (verbose >=1) print(parsed);
-          guids = parsed;//['message']['guids'];
-          if (verbose >=1) print('List of GUIDS readed. Updating UserInfo from Server');
+          if (verbose >= 1) print(parsed);
+          guids = parsed; //['message']['guids'];
+          if (verbose >= 1)
+            dprintL('List of GUIDS readed. Updating UserInfo from Server');
           currentGuidIndex = 0;
           //есть список гуидов и есть токен.
           //можно обновить данные с сервера или если не получится считать из файлов
           guidsNumber = guids.length;
           for (var guid in guids) {
-            if (verbose >=1) print('Updating from Server for guid $guid');
+            dprintL('Updating from Server for guid $guid');
             var result = await RestAPI().userDataGet(guid, devKey);
             if (result == 'Ok') {
-              if (verbose >=1) print('Ok. File $guid.dat and users[$currentGuidIndex] updated');
+              if (verbose >= 1)
+                dprintL(
+                    'Ok. File $guid.dat and users[$currentGuidIndex] updated');
             } else {
-              if (verbose >=1) print('Not ok!');
-              if (verbose >=1) print('Trying to read from file instead');
+              dprintL('Not ok!');
+              dprintL('Trying to read from file instead');
               await UserInfo().readFromFile();
             }
             currentGuidIndex++;
@@ -162,32 +158,39 @@ class _SplashWidgetState extends State<SplashWidget> {
           //загрузка пушей
           final shared = await SharedPreferences.getInstance();
           sharedPushes = shared.getStringList('pushes') ?? [];
-          print(sharedPushes);
-          sharedPushes.forEach((element) {pushes.add(jsonDecode(element));});
+          dprintD('List of saved shared pushed:', verbose);
+          sharedPushes.forEach((element) {
+            //pushes.add(jsonDecode(element));
+            dprintD(element, verbose);
+            dprintD('------------------', verbose);
+          });
           //pushes = await Pushes().loadPushesFromFile('pushes.dat');
           goGo(1);
         } else {
-          if (verbose >=1) print('GUIDS file not exists. Got to Login Screen');
-          goGo(0);//return 0;
+          dprintL('GUIDS file not exists. Got to Login Screen');
+          goGo(0); //return 0;
         }
       } else {
-        if (verbose >=1) print('Key File has wrong key. Trying to get new one from FireBase...');
+        dprintL(
+            'Key File has wrong key. Trying to get new one from FireBase...');
         String _token = await FirebaseHelper().getAppToken();
         devKey = _token;
-        if (verbose >=1) print(_token);
-        if (_token != null) file.writeAsStringSync(_token, mode: FileMode.write, encoding: utf8);
-        if (verbose >=1) print('New Key File created and saved');
-        if (verbose >=1) print('Go to Login Screen');
-        goGo(0);//return 0;
+        dprintL(_token);
+        if (_token != null)
+          file.writeAsStringSync(_token, mode: FileMode.write, encoding: utf8);
+        dprintL('New Key File created and saved');
+        dprintL('Go to Login Screen');
+        goGo(0); //return 0;
       }
     } else {
-      if (verbose >=1) print('Key file is not exists');
+      dprintL('Key file is not exists');
       String _token = await FirebaseHelper().getAppToken();
       devKey = _token;
-      if (verbose >=1) print(_token == null ? '' : _token);
-      file.writeAsStringSync(_token == null ? '' : _token, mode: FileMode.write, encoding: utf8);
-      if (verbose >=1) print('Key File created and saved');
-      if (verbose >=1) print('Go to Login Screen');
+      dprintL(_token == null ? '' : _token);
+      file.writeAsStringSync(_token == null ? '' : _token,
+          mode: FileMode.write, encoding: utf8);
+      dprintL('Key File created and saved');
+      dprintL('Go to Login Screen');
       goGo(0);
     }
   }
